fix: resolve 3 edge case bugs in raycaster, BSP, and quest systems
- Raycaster: Use far plane distance (1e10) instead of infinity (1e30) for horizon pixels to ensure consistent fog/lighting behavior
- BSP: Add nil/empty/bounds validation in placeSecrets() to prevent panics with invalid tile maps
- Quest: Change Progress field from int to int64 to prevent overflow in long gameplay sessions (tested up to 1 billion updates)

Added comprehensive unit tests for all edge cases:
- TestPlaceSecretsWithNilTiles: validates nil, empty, and undersized tile map handling
- TestPlaceSecretsWithValidTiles: ensures normal operation with valid tiles
- TestUpdateProgressInt64: tests 1 billion incremental updates
- TestUpdateProgressLargeValue: tests 100 million single update

All existing tests pass. No regressions.

Closes AUDIT.md edge case bugs #7, #8, #9.

+++ pkg/quest/quest.go
@@ -36,7 +36,7 @@ type Objective struct {
 	Desc     string
 	Target   string
 	Count    int
-	Progress int
+	Progress int64 // int64 prevents overflow in long gameplay sessions
 	Complete bool
 	PosX     float64 // Objective position in level
 	PosY     float64
@@ -235,8 +235,8 @@ func (t *Tracker) Add(obj Objective) {
 func (t *Tracker) UpdateProgress(id string, amount int) {
 	for i := range t.Objectives {
 		if t.Objectives[i].ID == id {
-			t.Objectives[i].Progress += amount
-			if t.Objectives[i].Progress >= t.Objectives[i].Count {
+			t.Objectives[i].Progress += int64(amount)
+			if t.Objectives[i].Progress >= int64(t.Objectives[i].Count) {
 				t.Objectives[i].Complete = true
 			}
 		}
